Main work completed
