+++ src/BannerNewsletter/index.stories.tsx
@@ -68,7 +68,7 @@ const StoryTemplate = (
         'https://i.guim.co.uk/img/media/568c6031be78dab6f6c28336010884f3ebd0f97c/0_0_1936_1936/master/1936.png?width=196&quality=45&auto=format&s=2a3630e9625620d5726c31c5cdbf4772',
     );
 
-    const brazeMessageProps: BrazeMessageProps = {
+    const brazeMessageProps = {
         header: args.header,
         newsletterId: args.newsletterId,
         frequency: args.frequency,
