+++ src/BannerNewsletter/index.tsx
@@ -1,46 +1,24 @@
 import React, { useState } from 'react';
-import { css } from '@emotion/react';
-import { from, headline } from '@guardian/source-foundations';
-import { Button, SvgCross } from '@guardian/source-react-components';
-
-import { useEscapeShortcut, OnCloseClick, CLOSE_BUTTON_ID } from '../bannerCommon/bannerActions';
 import type { NewsletterSubscribeCallback } from '../types/dcrTypes';
 import type { TrackClick } from '../utils/tracking';
 
 import {
     NewsletterCtaButton,
-    defaultNewsletterCtaButtonColors
+    defaultNewsletterCtaButtonColors,
 } from '../components/NewsletterCtaButton';
 import {
     NewsletterFrequencyBlock,
-    defaultNewsletterFrequencyColors
+    defaultNewsletterFrequencyColors,
 } from '../components/NewsletterFrequencyBlock';
-import { defaultBannerNewsletterColors } from '../StyleableBannerNewsletter';
-import { selfServeStyles } from '../styles/bannerCommon';
+import { BannerCloseButton, defaultBannerCloseButtonColors } from '../components/BannerCloseButton';
+import {
+    getBannerNewsletterStyles,
+    defaultBannerNewsletterBaseColors,
+} from '../StyleableBannerNewsletter';
 
 import { canRender, COMPONENT_NAME } from './canRender';
 
 export { COMPONENT_NAME };
-
-const localStyles = {
-    heading: css`
-        ${headline.small({ fontWeight: 'bold' })};
-        margin: 0;
-        max-width: 100%;
-
-        ${from.mobileLandscape} {
-            ${headline.small({ fontWeight: 'bold' })};
-        }
-
-        ${from.tablet} {
-            ${headline.medium({ fontWeight: 'bold' })};
-            max-width: 100%;
-        }
-    `,
-    bold: css`
-        font-weight: bold;
-    `,
-};
 
 export type BrazeMessageProps = {
     ophanComponentId?: string;
@@ -79,25 +57,12 @@ const BannerNewsletter: React.FC<Props> = (props: Props) => {
         trackClick,
     } = props;
 
-    const styles = selfServeStyles(props.brazeMessageProps, defaultBannerNewsletterColors);
+    const styles = getBannerNewsletterStyles(
+        props.brazeMessageProps,
+        defaultBannerNewsletterBaseColors,
+    );
 
     const [showBanner, setShowBanner] = useState(true);
-
-    const onCloseClick: OnCloseClick = (evt, internalButtonId) => {
-        evt.preventDefault();
-        onCloseAction(internalButtonId);
-    };
-
-    const onCloseAction = (internalButtonId: number): void => {
-        setShowBanner(false);
-        document.body.focus();
-        trackClick({
-            internalButtonId,
-            ophanComponentId: ophanComponentId as string,
-        });
-    };
-
-    useEscapeShortcut(() => onCloseAction(CLOSE_BUTTON_ID));
 
     if (!showBanner) {
         return null;
@@ -107,17 +72,17 @@ const BannerNewsletter: React.FC<Props> = (props: Props) => {
         <div css={styles.wrapper}>
             <div css={styles.contentContainer}>
                 <div css={styles.topLeftComponent}>
-                    <div css={localStyles.heading}>{header}</div>
+                    <div css={styles.heading}>{header}</div>
                     <NewsletterFrequencyBlock
                         frequency={frequency}
                         colors={defaultNewsletterFrequencyColors}
                     />
-                    <p css={styles.paragraph}>
-                        {body} {boldText && <span css={localStyles.bold}>{boldText}</span>}
-                    </p>
-                    {secondParagraph && (
-                        <p css={[styles.paragraph, styles.secondParagraph]}>{secondParagraph}</p>
-                    )}
+                    <div css={styles.paragraph}>
+                        <p>
+                            {body} {boldText && <span css={styles.highlight}>{boldText}</span>}
+                        </p>
+                        {secondParagraph && <p css={styles.secondParagraph}>{secondParagraph}</p>}
+                    </div>
                     <NewsletterCtaButton
                         subscribeToNewsletter={subscribeToNewsletter}
                         newsletterId={newsletterId as string}
@@ -131,20 +96,12 @@ const BannerNewsletter: React.FC<Props> = (props: Props) => {
                     <div css={styles.centeredImage}>
                         <img src={imageUrl} alt="" />
                     </div>
-                    <div css={styles.iconPanel}>
-                        <Button
-                            icon={<SvgCross />}
-                            hideLabel={true}
-                            cssOverrides={styles.closeButton}
-                            priority="tertiary"
-                            size="small"
-                            aria-label="Close"
-                            onClick={(e) => onCloseClick(e, CLOSE_BUTTON_ID)}
-                            tabIndex={0}
-                        >
-                            {' '}
-                        </Button>
-                    </div>
+                    <BannerCloseButton
+                        trackClick={trackClick}
+                        setShowBanner={setShowBanner}
+                        ophanComponentId={ophanComponentId}
+                        colors={defaultBannerCloseButtonColors}
+                    />
                 </div>
             </div>
         </div>
