+++ src/StyleableBannerWithLink/index.stories.tsx
@@ -163,7 +163,7 @@ const StoryTemplate = (args: BrazeMessageProps & { componentName: string }): Rea
         'https://i.guim.co.uk/img/media/35d403182e4b262d37385281b19b763ee6b32f6a/58_0_1743_1046/master/1743.png?width=930&quality=45&auto=format&s=9ecd82413fef9883c1e7a0df2bf6abb1',
     );
 
-    const brazeMessageProps: BrazeMessageProps = {
+    const brazeMessageProps = {
         styleBackground: args.styleBackground,
         header: args.header,
         styleHeader: args.styleHeader,
