+++ src/StyleableBannerWithLink/index.tsx
@@ -1,48 +1,38 @@
 import React, { useState } from 'react';
-import { Button, SvgCross } from '@guardian/source-react-components';
-import { useEscapeShortcut, OnCloseClick, CLOSE_BUTTON_ID } from '../bannerCommon/bannerActions';
-import { canRender, COMPONENT_NAME } from './canRender';
 
 import { PrimaryCtaButton, defaultPrimaryCtaButtonColors } from '../components/PrimaryCtaButton';
 import { ReminderCtaButton, defaultReminderCtaButtonColors } from '../components/ReminderCtaButton';
-import { selfServeStyles } from '../styles/bannerCommon';
-import { getColors } from '../styles/colorData';
+import { BannerCloseButton, defaultBannerCloseButtonColors } from '../components/BannerCloseButton';
 
 import type { Extras } from '../logic/types';
 import type { ReminderStage } from '../logic/reminders';
 import type { TrackClick } from '../utils/tracking';
 import type { FetchEmail } from '../types/dcrTypes';
-import type { StyleableBannerWithLinkColorStyles, BannerWithLinkCopyColorStyles } from '../styles/colorData';
+import type {
+    StyleableBannerWithLinkColorStyles,
+    BannerWithLinkBaseColorStyles,
+} from '../styles/colorData';
 
+import { canRender, COMPONENT_NAME } from './canRender';
 export { COMPONENT_NAME };
 
-export const defaultBannerWithLinkColors: BannerWithLinkCopyColorStyles = {
-    styleBackground: '#ededed',
-    styleHeader: '#333333',
-    styleBody: '#333333',
-    styleHighlight: '#333333',
-    styleHighlightBackground: '#ededed',
-    styleClose: '#052962',
-    styleCloseBackground: '#ededed',
-    styleCloseHover: '#e5e5e5',
-};
-
-export type BrazeMessageProps = Extras & StyleableBannerWithLinkColorStyles & {
-    ophanComponentId?: string;
-    header?: string;
-    body?: string;
-    highlight?: string;
-    buttonText?: string;
-    buttonUrl?: string;
-    showPaymentIcons?: string;
-    reminderStage?: ReminderStage;
-    reminderOption?: string;
-    includeReminderCta?: string;
-    imageUrl?: string;
-    imageAltText?: string;
-    imagePosition?: string;
-    showPrivacyText?: string;
-};
+export type BrazeMessageProps = Extras &
+    StyleableBannerWithLinkColorStyles & {
+        ophanComponentId?: string;
+        header?: string;
+        body?: string;
+        highlight?: string;
+        buttonText?: string;
+        buttonUrl?: string;
+        showPaymentIcons?: string;
+        reminderStage?: ReminderStage;
+        reminderOption?: string;
+        includeReminderCta?: string;
+        imageUrl?: string;
+        imageAltText?: string;
+        imagePosition?: string;
+        showPrivacyText?: string;
+    };
 
 export type Props = {
     brazeMessageProps: BrazeMessageProps;
@@ -77,33 +67,16 @@ const StyleableBannerWithLink: React.FC<Props> = (props: Props) => {
 
     const brazeProps = props.brazeMessageProps;
 
+    const bannerWithLinkBaseColors = getColors(brazeProps, defaultBannerWithLinkColors);
     const primaryCtaStyles = getColors(brazeProps, defaultPrimaryCtaButtonColors);
     const reminderCtaStyles = getColors(brazeProps, defaultReminderCtaButtonColors);
+    const bannerCloseButtonColors = getColors(brazeProps, defaultBannerCloseButtonColors);
 
-    const styles = selfServeStyles(brazeProps, {
-        ...defaultBannerWithLinkColors,
-        ...defaultPrimaryCtaButtonColors,
-    });
+    const styles = getBannerWithLinkStyles(brazeProps, bannerWithLinkBaseColors);
 
     const showPrivacyTextBoolean = showPrivacyText === 'true';
 
     const [showBanner, setShowBanner] = useState(true);
-
-    const onCloseClick: OnCloseClick = (evt, internalButtonId) => {
-        evt.preventDefault();
-        onCloseAction(internalButtonId);
-    };
-
-    const onCloseAction = (internalButtonId: number): void => {
-        setShowBanner(false);
-        document.body.focus();
-        trackClick({
-            internalButtonId,
-            ophanComponentId: ophanComponentId as string,
-        });
-    };
-
-    useEscapeShortcut(() => onCloseAction(CLOSE_BUTTON_ID));
 
     if (!showBanner) {
         return null;
@@ -155,20 +128,12 @@ const StyleableBannerWithLink: React.FC<Props> = (props: Props) => {
                     <div css={imagePosition === 'bottom' ? styles.image : styles.centeredImage}>
                         <img src={imageUrl} alt={imageAltText} />
                     </div>
-                    <div css={styles.iconPanel}>
-                        <Button
-                            icon={<SvgCross />}
-                            hideLabel={true}
-                            cssOverrides={styles.closeButton}
-                            priority="tertiary"
-                            size="small"
-                            aria-label="Close"
-                            onClick={(e) => onCloseClick(e, CLOSE_BUTTON_ID)}
-                            tabIndex={0}
-                        >
-                            {' '}
-                        </Button>
-                    </div>
+                    <BannerCloseButton
+                        trackClick={trackClick}
+                        setShowBanner={setShowBanner}
+                        ophanComponentId={ophanComponentId}
+                        colors={bannerCloseButtonColors}
+                    />
                 </div>
             </div>
         </div>
@@ -176,3 +141,277 @@ const StyleableBannerWithLink: React.FC<Props> = (props: Props) => {
 };
 
 export { StyleableBannerWithLink };
+
+// Styling
+// --------------------------------------------
+import { css } from '@emotion/react';
+import { neutral, space, from, until, body, headline } from '@guardian/source-foundations';
+import { getColors } from '../styles/colorData';
+
+export const defaultBannerWithLinkColors: BannerWithLinkBaseColorStyles = {
+    styleBackground: '#ededed',
+    styleHeader: '#333333',
+    styleBody: '#333333',
+    styleHighlight: '#333333',
+    styleHighlightBackground: '#ededed',
+};
+
+const imgHeight = '280';
+
+export const getBannerWithLinkStyles = (
+    userVals: Record<string, string>,
+    defaults: BannerWithLinkBaseColorStyles,
+) => {
+    const colors = getColors(userVals, defaults);
+
+    return {
+        wrapper: css`
+            box-sizing: border-box;
+            display: flex;
+            flex-direction: column;
+            justify-content: center;
+            align-items: center;
+            width: 100%;
+            background-color: ${colors.styleBackground};
+            color: ${neutral[20]};
+
+            html {
+                box-sizing: border-box;
+            }
+
+            *,
+            *:before,
+            *:after {
+                box-sizing: inherit;
+            }
+        `,
+
+        contentContainer: css`
+            align-items: stretch;
+            display: flex;
+            flex-direction: column;
+            position: relative;
+            margin: 0 auto;
+            width: 100%;
+            max-width: 740px;
+
+            ${from.desktop} {
+                max-width: 980px;
+                flex-direction: row;
+                min-height: ${imgHeight}px;
+            }
+
+            ${from.leftCol} {
+                max-width: 1140px;
+            }
+
+            ${from.wide} {
+                max-width: 1300px;
+            }
+        `,
+
+        topLeftComponent: css`
+            width: 93%;
+            padding: ${space[4]}px;
+            min-height: ${imgHeight}px;
+            ${from.desktop} {
+                width: 60%;
+            }
+        `,
+
+        bottomRightComponent: css`
+            display: flex;
+            justify-content: center;
+            width: 100%;
+            max-height: 100%;
+
+            ${from.desktop} {
+                align-self: flex-end;
+                padding-right: ${space[4]}px;
+                max-width: 45%;
+                justify-content: flex-end;
+            }
+
+            ${from.leftCol} {
+                padding-right: 0;
+                justify-content: space-between;
+            }
+
+            ${from.wide} {
+                max-width: 48%;
+            }
+        `,
+
+        centeredBottomRightComponent: css`
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            width: 100%;
+            max-height: 100%;
+
+            ${from.desktop} {
+                padding-right: ${space[4]}px;
+                max-width: 45%;
+            }
+
+            ${from.leftCol} {
+                padding-right: 0;
+            }
+
+            ${from.wide} {
+                max-width: 48%;
+            }
+        `,
+
+        heading: css`
+            ${headline.small({ fontWeight: 'bold' })};
+            margin: 0;
+            max-width: 100%;
+            color: ${colors.styleHeader};
+
+            ${from.mobileLandscape} {
+                ${headline.small({ fontWeight: 'bold' })};
+            }
+
+            ${from.tablet} {
+                max-width: 100%;
+            }
+        `,
+
+        paragraph: css`
+            ${body.medium()}
+            line-height: 135%;
+            margin: ${space[5]}px 0 ${space[5]}px;
+            max-width: 100%;
+            color: ${colors.styleBody};
+
+            ${from.phablet} {
+                max-width: 90%;
+            }
+
+            ${from.tablet} {
+                max-width: 100%;
+            }
+
+            ${from.desktop} {
+                font-size: 20px;
+                margin: ${space[3]}px 0 ${space[4]}px;
+                max-width: 42rem;
+            }
+
+            ${from.leftCol} {
+                max-width: 37rem;
+            }
+
+            ${from.wide} {
+                max-width: 42rem;
+            }
+        `,
+
+        highlightContainer: css`
+            ${body.medium()}
+            margin-top: ${space[5]}px;
+            margin-right: ${space[3]}px;
+            max-width: 100%;
+
+            ${from.phablet} {
+                max-width: 90%;
+            }
+
+            ${from.tablet} {
+                max-width: 100%;
+            }
+
+            ${from.desktop} {
+                font-size: 20px;
+                margin: ${space[3]}px 0 ${space[4]}px;
+                max-width: 42rem;
+            }
+
+            ${from.leftCol} {
+                max-width: 37rem;
+            }
+
+            ${from.wide} {
+                max-width: 42rem;
+            }
+        `,
+
+        highlight: css`
+            font-weight: 700;
+            color: ${colors.styleHighlight};
+            background-color: ${colors.styleHighlightBackground};
+        `,
+
+        smallRightSpacer: css`
+            margin-right: ${space[3]}px;
+        `,
+
+        primaryButtonWrapper: css`
+            margin: 0 ${space[2]}px ${space[1]}px 0;
+            display: flex;
+            flex-wrap: wrap;
+            justify-content: flex-start;
+            align-items: flex-start;
+        `,
+
+        iconPanel: css`
+            ${from.desktop} {
+                display: flex;
+                flex-direction: column;
+                justify-content: space-between;
+                align-items: flex-end;
+                height: 100%;
+                padding: ${space[4]}px 0;
+                margin-left: ${space[4]}px;
+            }
+        `,
+
+        image: css`
+            max-width: 100%;
+            max-height: 300px;
+            display: flex;
+            justify-content: center;
+            align-items: flex-end;
+            margin-top: ${space[2]}px;
+
+            img {
+                max-width: 100%;
+                width: 100%;
+                height: 100%;
+                object-fit: contain;
+            }
+
+            ${until.desktop} {
+                display: none;
+            }
+
+            ${from.wide} {
+                margin-right: 100px;
+            }
+        `,
+        centeredImage: css`
+            max-width: 100%;
+            max-height: 300px;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            margin-top: ${space[2]}px;
+
+            img {
+                max-width: 100%;
+                width: 100%;
+                height: 100%;
+                object-fit: contain;
+            }
+
+            ${until.desktop} {
+                display: none;
+            }
+
+            ${from.wide} {
+                margin-right: 100px;
+            }
+        `,
+    };
+};
